+++ docker/dedicated/updater.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
 cd /tmp/
-wget http://files.v04.maniaplanet.com/server/ManiaplanetServer_Latest.zip
-unzip ManiaplanetServer_Latest.zip
-mv ManiaPlanetServer /home/maniaplanet/
+wget http://files.v04.maniaplanet.com/server/TrackmaniaServer_Latest.zip
+unzip TrackmaniaServer_Latest.zip
+mv TrackmaniaServer_Latest /home/trackmania/TrackmaniaServer