+++ src/event.rs
@@ -1,9 +1,8 @@
 use std::collections::HashMap;
 
-use gbx::CheckpointEvent;
-
 use crate::chat::{AdminCommand, PlayerCommand, SuperAdminCommand};
-use crate::database::{Map, RecordDetailed};
+use crate::database::{Map, Record};
+use crate::server::CheckpointEvent;
 use crate::server::{GameString, PlayerInfo};
 use crate::widget::Action;
 
@@ -180,8 +179,8 @@ pub struct PbDiff {
     /// The player's map rank before this run.
     pub prev_pos: Option<usize>,
 
-    /// The player's map rank after this run, which has improved
-    /// if `new_record` is `Some`, and is equal to `prev_pos` otherwise.
+    /// The player's map rank after this run, which may have improved
+    /// if `new_record` is `Some`.
     pub new_pos: usize,
 
     /// If `prev_pos` is `Some`, this is `prev_pos - new_pos`.
@@ -189,7 +188,7 @@ pub struct PbDiff {
     pub pos_gained: usize,
 
     /// `Some` if the player improved their personal best.
-    pub new_record: Option<RecordDetailed>,
+    pub new_record: Option<Record>,
 }
 
 /// A command with its sender, who was confirmed to have the necessary permission.