+++ src/server.rs
@@ -1,1 +1,1 @@
-pub use gbx::{Callback as ServerEvent, Calls as Server, *};
+pub use gbx::{file as _, Callback as ServerEvent, Calls as Server, *};